Pass the user object when logging in a user.
We used to set the current user on the context object before logging in the
user. This caused the state of the user to be inconsistent. Now we pass in
the user object to the methods responsible for logging in, and the state is
only updated when the user is successfully logged in.