correct paths per request
How's this?  

Behat is new to me, so I have only knowledge of what I encountered when I attempted to follow the installation instructions.  Wasn't sure what to do about the file path in this text: _"The `features/FeatureContext.php` file must be in the same directory as your `behat.yml`"_