Update RawDrupalContext.php
With Scenarios there the first step has to do with login/logout I get the following error: 

Given I am logged in as a user with the "XXXX" role # Drupal\DrupalExtension\Context\DrupalContext::assertAuthenticatedByRole()
      Fatal error: Call to a member function elements() on null (Behat\Testwork\Call\Exception\FatalThrowableError)

Because the Selenium2Driver is not startet at this moment. With this fix in loggedIn() it works.