remove sudo from composer command
Composer maintainers strongly advise against running composer as sudo/root. On Mac and Ubuntu it appears not to be necessary. Is it possible to accomplish everything without using sudo with composer?

See their FAQ:
How do I install untrusted packages safely? Is it safe to run Composer as superuser or root?
https://getcomposer.org/doc/faqs/how-to-install-untrusted-packages-safely.md